Importants improvements and changes:
ReceivingAndMaekingTransfer_SouldHaveValidTransactionChain should be ReceivingAndMaekingTransfer_SouldHaveValidTransaction because its validation the transaction and not the transaction chain

The transaction.checktransaction should validate just yourself and not the transaction chain. The person should validate the entiery chain.

The validation isFirstTransaction was not object oriented. Was created GoogyTransactoin which represents the first transaction and inherit transaction with checktransaction virtual method so than the firsttransaction and transactions validations are made as the virtual method